TruncateFieldUpdateProcessorFactory maxLength=8000
Reduce TruncateFieldUpdateProcessorFactory maxLength to 8000. This value is a character count and UTF8 can be 1-4 bytes. 8000 ensures there are less than 32000 bytes. Note that it is very unlikely that all the characters would be 4 bytes so a higher length would probably suffice in all be the very edgiest of cases. In fact 3 bytes covers all current languages.